+++ srtcore/buffer_tools.h
@@ -131,61 +131,61 @@ class CSndRateEstimator
 {
     typedef sync::steady_clock::time_point time_point;
 public:
-	CSndRateEstimator(const time_point& tsNow);
+    CSndRateEstimator(const time_point& tsNow);
 
     /// Add sample.
     /// @param [in] time   sample (sending) time.
     /// @param [in] pkts   number of packets in the sample.
     /// @param [in] bytes  number of payload bytes in the sample.
-	void addSample(const time_point& time, int pkts = 0, int bytes = 0);
+    void addSample(const time_point& time, int pkts = 0, size_t bytes = 0);
 
     /// Retrieve input bitrate in bytes per second
     int getRate() const { return m_iRateBps; }
 
 private:
-	static const int NUM_PERIODS = 10;
-	static const int SAMPLE_DURATION_MS = 100;  // 100 ms
+    static const int NUM_PERIODS = 10;
+    static const int SAMPLE_DURATION_MS = 100;  // 100 ms
     struct Sample
     {
         int m_iPktsCount;  // number of payload packets
         int m_iBytesCount; // number of payload bytes
 
-		void reset()
-		{
-			m_iPktsCount = 0;
-			m_iBytesCount = 0;
-		}
+        void reset()
+        {
+            m_iPktsCount = 0;
+            m_iBytesCount = 0;
+        }
 
-		Sample()
-			: m_iPktsCount(0)
-			, m_iBytesCount(0)
-		{ }
+        Sample()
+            : m_iPktsCount(0)
+            , m_iBytesCount(0)
+        { }
 
-		Sample(int iPkts, int iBytes)
-			: m_iPktsCount(iPkts)
-			, m_iBytesCount(iBytes)
-		{ }
+        Sample(int iPkts, int iBytes)
+            : m_iPktsCount(iPkts)
+            , m_iBytesCount(iBytes)
+        { }
 
-		Sample operator+(const Sample& other)
-		{
-			return Sample(m_iPktsCount + other.m_iPktsCount, m_iBytesCount + other.m_iBytesCount );
-		}
+        Sample operator+(const Sample& other)
+        {
+            return Sample(m_iPktsCount + other.m_iPktsCount, m_iBytesCount + other.m_iBytesCount );
+        }
 
-		Sample& operator+=(const Sample& other)
-		{
-			*this = *this + other;
-			return *this;
-		}
+        Sample& operator+=(const Sample& other)
+        {
+            *this = *this + other;
+            return *this;
+        }
 
-		bool empty() const
-		{
-			return m_iPktsCount == 0;
-		}
+        bool empty() const
+        {
+            return m_iPktsCount == 0;
+        }
     };
 
-	int incSampleIdx(int val, int inc = 1) const;
+    int incSampleIdx(int val, int inc = 1) const;
 
-	Sample m_Samples[NUM_PERIODS] = {};
+    Sample m_Samples[NUM_PERIODS] = {};
 
     time_point m_tsFirstSampleTime; //< Start time of the first sameple.
     int m_iFirstSampleIdx = 0; //< Index of the first sample.
